Log errors when mounting views to the console for easier reference
From the console, I can click the source references in the stack trace
to jump to the spot in the code where things are going wrong.

It also helps with the problem of the error not having enough
space to be read in some components.

Split off from https://github.com/vector-im/hydrogen-web/pull/653