Load features in platform init instead of main
So that features are available through the Platform. I think logically it makes sense to have the Platform load the features, as where features are coming from depends on the Platform. On Web they come from localStorage, on other platforms they might come from elsewhere.

This will allow us to inject the FeatureSet into a SyncFactory class in next commits.